requirements.txt
asgiref==3.7.2
Django==4.2.2
pip==23.1.2
setuptools==68.0.0
sqlparse==0.4.4
tzdata==2023.3
djangorestframework==3.14.0
django-environ==0.10.0
psycopg2-binary==2.9.6
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
django-cors-headers==4.1.0
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0